Break down classes and add Settings class
