+++ models/dictionary.py
@@ -2,71 +2,7 @@ import os
 import csv
 
 from utils.search import bisect_left
-
-
-class Cluster:
-    """
-    The basic unit of definition(s) and info of a word.
-    """
-
-    def __init__(self, meanings: list[str] = [], examples: list[str] = [], synonyms: list[str] = [], antonyms: list[str] = [], related: list[str] = []) -> None:
-        self.meanings = meanings
-        self.examples = examples
-        self.synonyms = synonyms
-        self.antonyms = antonyms
-        self.related = related
-
-    def add_meaning(self, meaning: str) -> None:
-        self.meanings.append(meaning)
-
-    def add_example(self, example: str) -> None:
-        self.examples.append(example)
-
-    def add_synonym(self, synonym: str) -> None:
-        self.synonyms.append(synonym)
-
-    def add_antonym(self, antonym: str) -> None:
-        self.antonyms.append(antonym)
-
-    def add_related(self, related: str) -> None:
-        self.related.append(related)
-
-
-class Vocabulary:
-    """
-    A word defined by Cluster(s).
-    """
-
-    def __init__(self, word: str = "") -> None:
-        self.word = word
-        self.clusters = {}
-
-    def __gt__(self, other) -> bool:
-        return self.word > other.word
-
-    def __lt__(self, other) -> bool:
-        return self.word < other.word
-
-    def __eq__(self, other) -> bool:
-        if not isinstance(other, Vocabulary):
-            return False
-
-        return self.word == other.word
-
-    def __hash__(self) -> int:
-        return hash(self.word)
-
-    def __repr__(self) -> str:
-        return self.word
-
-    def add_cluster(self, pos: str, cluster: Cluster) -> None:
-        self.clusters[pos] = cluster
-
-    def get_cluster(self, pos: str) -> Cluster | None:
-        return self.clusters.get(pos, None)
-
-    def get_size(self) -> int:
-        return len(self.clusters)
+from models.vocabulary import Cluster, Vocabulary
 
 
 class Dictionary:
@@ -85,9 +21,6 @@ class Dictionary:
 
     def remove_word(self, word: str) -> None:
         del self.vocabs[word]
-
-    def clear(self) -> None:
-        self.vocabs.clear()
 
     def get_vocab(self, word: str) -> Vocabulary | None:
         return self.vocabs.get(word, None)
@@ -111,6 +44,9 @@ class Dictionary:
 
     def sort(self) -> None:
         self.vocabs = dict(sorted(self.vocabs.items()))
+
+    def reset(self) -> None:
+        self.vocabs.clear()
 
     def to_csv(self, path: str, delim: str = "|") -> None:
         with open(path, "w+", newline="") as f:
